`tools/generator-go-sdk`: splitting out the existing LRO tests
