+++ tools/generator-go-sdk/generator/templater_methods_test.go
@@ -153,181 +153,6 @@ func (c pandaClient) Get(ctx context.Context , id PandaPop) (result GetOperation
 	assertTemplatedCodeMatches(t, expected, *actual)
 }
 
-func TestTemplateMethodsLROCreate(t *testing.T) {
-	input := ServiceGeneratorData{
-		packageName:       "skinnyPandas",
-		serviceClientName: "pandaClient",
-		source:            AccTestLicenceType,
-		resourceIds: map[string]resourcemanager.ResourceIdDefinition{
-			"PandaPop": {
-				Id: "LingLing",
-			},
-		},
-	}
-
-	actual, err := methodsPandoraTemplater{
-		operation: resourcemanager.ApiOperation{
-			ExpectedStatusCodes: []int{201, 202},
-			LongRunning:         true,
-			Method:              "PUT",
-			RequestObject: &resourcemanager.ApiObjectDefinition{
-				Type: resourcemanager.StringApiObjectDefinitionType,
-			},
-			ResourceIdName: stringPointer("PandaPop"),
-		},
-		operationName: "Create",
-	}.longRunningOperationTemplate(input)
-
-	if err != nil {
-		t.Fatalf("err %+v", err)
-	}
-
-	expected := `
-type CreateOperationResponse struct {
-	Poller pollers.Poller
-	HttpResponse *http.Response
-	OData *odata.OData
-}
-
-// Create ...
-func (c pandaClient) Create(ctx context.Context , id PandaPop, input string) (result CreateOperationResponse, err error) {
-	opts := client.RequestOptions{
-		ContentType: "application/json",
-		ExpectedStatusCodes: []int{
-			http.StatusAccepted,
-			http.StatusCreated,
-		},
-		HttpMethod: http.MethodPut,
-		Path: id.ID(),
-	}
-
-	req, err := c.Client.NewRequest(ctx, opts)
-	if err != nil {
-		return
-	}
-
-	if err = req.Marshal(input); err != nil {
-		return
-	}
-
-	var resp *client.Response
-	resp, err = req.Execute(ctx)
-	if resp != nil {
-		result.OData = resp.OData
-		result.HttpResponse = resp.Response
-	}
-	if err != nil {
-		return
-	}
-
-	result.Poller, err = resourcemanager.PollerFromResponse(resp, c.Client)
-	if err != nil {
-		return
-	}
-
-	return
-}
-
-// CreateThenPoll performs Create then polls until it's completed
-func (c pandaClient) CreateThenPoll(ctx context.Context , id PandaPop, input string) error {
-	result, err := c.Create(ctx , id, input)
-	if err != nil {
-		return fmt.Errorf("performing Create: %+v", err)
-	}
-
-	if err := result.Poller.PollUntilDone(ctx); err != nil {
-		return fmt.Errorf("polling after Create: %+v", err)
-	}
-
-	return nil
-}
-`
-	assertTemplatedCodeMatches(t, expected, *actual)
-}
-
-func TestTemplateMethodsLROReboot(t *testing.T) {
-	input := ServiceGeneratorData{
-		packageName:       "skinnyPandas",
-		serviceClientName: "pandaClient",
-		source:            AccTestLicenceType,
-		resourceIds: map[string]resourcemanager.ResourceIdDefinition{
-			"PandaPop": {
-				Id: "LingLing",
-			},
-		},
-	}
-
-	actual, err := methodsPandoraTemplater{
-		operation: resourcemanager.ApiOperation{
-			ExpectedStatusCodes: []int{200},
-			LongRunning:         true,
-			Method:              "POST",
-			ResourceIdName:      stringPointer("PandaPop"),
-		},
-		operationName: "Reboot",
-	}.longRunningOperationTemplate(input)
-
-	if err != nil {
-		t.Fatalf("err %+v", err)
-	}
-
-	expected := `
-type RebootOperationResponse struct {
-	Poller pollers.Poller
-	HttpResponse *http.Response
-	OData *odata.OData
-}
-
-// Reboot ...
-func (c pandaClient) Reboot(ctx context.Context , id PandaPop) (result RebootOperationResponse, err error) {
-	opts := client.RequestOptions{
-		ContentType: "application/json",
-		ExpectedStatusCodes: []int{
-			http.StatusOK,
-		},
-		HttpMethod: http.MethodPost,
-		Path: id.ID(),
-	}
-
-	req, err := c.Client.NewRequest(ctx, opts)
-	if err != nil {
-		return
-	}
-	var resp *client.Response
-	resp, err = req.Execute(ctx)
-	if resp != nil {
-		result.OData = resp.OData
-		result.HttpResponse = resp.Response
-	}
-	if err != nil {
-	return
-	}
-
-	result.Poller, err = resourcemanager.PollerFromResponse(resp, c.Client)
-	if err != nil {
-		return
-	}
-
-	return
-}
-
-// RebootThenPoll performs Reboot then polls until it's completed
-func (c pandaClient) RebootThenPoll(ctx context.Context , id PandaPop) error {
-	result, err := c.Reboot(ctx , id)
-	if err != nil {
-		return fmt.Errorf("performing Reboot: %+v", err)
-	}
-
-	if err := result.Poller.PollUntilDone(ctx); err != nil {
-		return fmt.Errorf("polling after Reboot: %+v", err)
-	}
-
-	return nil
-}
-`
-	assertTemplatedCodeMatches(t, expected, *actual)
-}
-
 // NOTE: List Operations differ slightly depending on the Response Object.
 // * When the Response Object is a Model or a List (of a Model) we include support for predicate filtering.
 // * When the Response Object is a Simple Type (e.g. int/string) we don't.
